Home work 22 JUnit
