+++ src/test/java/home_work_22/test/junit/ArraysTest.java
@@ -1,4 +1,4 @@
-package home_work_22.test;
+package home_work_22.test.junit;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,6 +11,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
+import static home_work_22.test.junit.ArraysUtl.*;
 import static home_work_22.test.junit.Main.*;
